services/orchestrator/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
